user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["duciito <duciito@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
uvicorn = {version = "^0.25.0", extras = ["standard"]}
motor = "3.3.2"
beanie = "1.24.0"
email-validator = "2.1.0.post1"
fastapi-jwt-auth = {version = "0.5.0", extras = ["asymmetric"]}
fastapi-mail = "1.4.1"
aioboto3 = "^12.1.0"
redis = "5.0.1"
websockets = "^12.0"
pydantic-settings = "^2.1.0"

[tool.ruff]
select = [ "E", "F", "UP", "ISC", "I", "W", "B" ]
target-version = "py311"

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Body", "fastapi.Query", "fastapi.params.Query"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
